Finished project

+++ script.ps1
@@ -1,10 +1,3 @@
-#Überprüft ob das Modul bereits installiert wurde.
-if (!(Get-Module -ListAvailable -Name Mailozaurr)) {
-    #Wenn das Modul nicht installiert ist, installiere das Modul
-    Install-Module -Name Mailozaurr -AllowPrerelease2
-}
-
-# Welcome
 Write-Host "|"
 Write-Host "|  Willkommen zum CLI E-Mail Tool"
 Write-Host "--------------------------------------------------------------------------------------"
@@ -13,49 +6,76 @@ Write-Host "|  © 2024 by Levyn Schneider and David Meer"
 Write-Host "| (Unterstützt nur Gmail)"
 Write-Host "|"
 
-# Read Sender E-Mail from txt file
-$email = $null
-
-# Simple function to create a new file and insert the email
-function Get-Email() {
+# Function to handle email and app-code input
+function Get-EmailAndCode() {
   Param($text)
 
-  $email = Read-Host "|  $text"
-
-  if ($email -match "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$" -and $email.ToLower().Contains("@gmail.com")) {
-    $email | Out-File .\settings.txt
-  } else {
-    Get-Email "Diese E-Mail-Adresse ist ungültig. Bitte geben Sie eine gültige E-Mail-Adresse ein"
+  # Delete settings.txt if it exists to avoid conflicts
+  if (Test-Path .\settings.txt) {
+    Remove-Item .\settings.txt
   }
 
-  $password = Read-Host "|  Bitte geben Sie Ihr E-Mail App Passwort von Google ein"
-  $password | Out-File .\settings.txt -Append
+    Write-Host "|  $text"
+    $sender = $null
 
-  $content = Get-Content .\settings.txt
-  Write-Host "|"
-  Write-Host "--------------------------------------------------------------------------------------" + $content[1]
-  Write-Host ""
+    while ($sender -eq $null) {
+      $sender = Read-Host "|  Bitte geben Sie Ihre E-Mail Adresse ein"
+      if ($sender -match "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$" -and $sender -like "*@gmail.com") {
+        $sender | Out-File .\settings.txt
+
+        $password = $null
+
+        while ($password -eq $null) {
+          $password = Read-Host "|  Bitte geben Sie Ihren Gmail App-Code ein"
+
+          if ($password -eq "") {
+            Write-Host "|"
+            Write-Host "|  Der Gmail App-Code darf nicht leer sein"
+            $password = $null
+            continue
+          }
+
+          $password | Out-File .\settings.txt -Append
+          Write-Host "|"
+          Write-Host "|  E-Mail Adresse und Gmail App-Code erfolgreich konfiguriert"
+          Write-Host "|"
+          Write-Host "--------------------------------------------------------------------------------------"
+          Write-Host
+        }
+      } else {
+        Write-Host "|"
+        Write-Host "|  Diese E-Mail Adresse ist ungültig. Bitte geben Sie eine gültige E-Mail Adresse ein."
+        $sender = $null
+      }
+    }
 }
 
 # Check if file exists
 if (-not (Test-Path .\settings.txt)) {
     New-Item -ItemType file -Path .\settings.txt | Out-Null
-    Get-Email "Bitte geben Sie Ihre E-Mail Adresse ein um fortzufahren"
+    Get-EmailAndCode "Bitte geben Sie Ihre E-Mail Adresse und Gmail App Code ein um fortzufahren"
 } else {
-  $content = Get-Content .\settings.txt
-  $email = $content[0]
+  $Content = Get-Content .\settings.txt
 
-  if ($email -ne "" || $email -ne $null) {
-    if ($email -match "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$") {
-      Write-Host "|  Konfigurierte Sender E-Mail Adresse: $email"
+  # Check if the content has two lines
+  if ($Content.Count -ne 2) {
+    Get-EmailAndCode "Falsche Konfiguration; Bitte geben Sie Ihre E-Mail Adresse und Gmail App Code ein um fortzufahren"
+  }
+
+  $Email = $Content[0]
+  $Password = $Content[1]
+
+  if ($Email -ne "" -and $Password -ne "") {
+    if ($Email -match "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$" -and $Email -like "*@gmail.com") {
+      Write-Host "|  Konfigurierte Sender E-Mail Adresse: $Email"
       Write-Host "|"
       Write-Host "--------------------------------------------------------------------------------------"
-      Write-Host ""
+      Write-Host
     } else {
-      Get-Email "Deine Konfigurierte E-Mail Adresse ist ungültig. Bitte gib eine gültige E-Mail Adresse ein"
+      Get-EmailAndCode "Deine Konfigurierte E-Mail Adresse ist ungültig. Bitte gib eine gültige E-Mail Adresse ein"
     }
   } else {
-    Get-Email "Bitte geben Sie Ihre E-Mail Adresse ein um fortzufahren"
+    Get-EmailAndCode "Bitte geben Sie Ihre E-Mail Adresse und Gmail App Code ein um fortzufahren"
   }
 }
 
@@ -65,9 +85,8 @@ Write-Host "1 (Sender E-Mail Einstellungen)"
 Write-Host "2 (E-Mail Senden)"
 Write-Host "3 (E-Mail Empfangen)"
 Write-Host "4 (Exit)"
+Write-Host
 
-# User Input
-Write-Host ""
 $option = Read-Host "Option wählen"
 
 switch ($option) {
@@ -75,7 +94,7 @@ switch ($option) {
     2 { .\send.ps1 }
     3 { .\receive.ps1 }
     4 { 
-      Write-Host ""
+      Write-Host
       Write-Host "Auf Wiedersehen!"
      }
     default { Write-Host "Ungültige Option" }
